refactored folder contents; pulled jsLDA.js from index/html; reformatted and reorganized index.html
