--- VARAIBLES\-DOC.md
@@ -1,2 +1,0 @@
-##  jsLDA Variable list
-- 2016-03-27
