+++ data/stop_words.txt
@@ -1,0 +1,65 @@
+the
+and
+of
+for
+in
+a
+on
+is
+an
+this
+to
+by
+our
+that
+will
+have
+are
+with
+all
+must
+not
+more
+their
+has
+but
+can
+people
+new
+world
+from
+congress
+year
+which
+they
+these
+you
+years
+now
+than
+been
+who
+should
+its
+one
+make
+every
+other
+those
+them
+time
+was
+also
+there
+many
+great
+last
+first
+only
+would
+when
+most
+need
+own
+what
+because
