--- out.gv
@@ -1,81 +1,0 @@
-digraph G {
-concentrate = true;
-			subgraph legend{
-			rank = min;
-			label = "legend";
-			Legend [shape=none, margin=0, label = <
-				<table cellspacing="0" cellpadding="0" border="1"><tr><td>Code2flow Legend</td></tr><tr><td>
-				<table cellspacing="0">
-				<tr><td>Regular function</td><td width="50px"></td></tr>
-				<tr><td>Trunk function (nothing calls this)</td><td bgcolor='coral'></td></tr>
-				<tr><td>Leaf function (this calls nothing else)</td><td bgcolor='green'></td></tr>
-				<tr><td>Function call which returns no value</td><td>&#8594;</td></tr>
-				<tr><td>Function call returns some value</td><td><font color='blue'>&#8594;</font></td></tr>
-				</table></td></tr></table>
-				>];}node100 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "0: (jsLDA window frame (runs on import))" ];
-node3 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "2: QueryString" ];
-node5 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "34: zeros" ];
-node10 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "104: truncate" ];
-node14 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "111: parseLine" ];
-node16 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "168: sampleDiscrete" ];
-node20 [splines=ortho shape = "rect" style = "rounded" label = "179: sweep" ];
-node22 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "229: byCountDescending" ];
-node26 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "232: topNWords" ];
-node28 [splines=ortho shape = "rect" style = "rounded" label = "238: sortTopicWords" ];
-node32 [splines=ortho shape = "rect" style = "rounded" label = "258: displayTopicWords" ];
-node34 [splines=ortho shape = "rect" style = "rounded" label = "282: reorderDocuments" ];
-node38 [splines=ortho shape = "rect" style = "rounded" label = "330: getTopicCorrelations" ];
-node41 [splines=ortho shape = "rect" style = "rounded" label = "377: getCorrelationGraph" ];
-node47 [splines=ortho shape = "rect" style = "rounded" label = "405: plotMatrix" ];
-node49 [splines=ortho shape = "rect" style = "rounded" label = "478: toggleTopicDocuments" ];
-node51 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "494: mostFrequentWords" ];
-node55 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "green" label = "508: entropy" ];
-node61 [splines=ortho shape = "rect" style = "rounded" label = "516: vocabTable" ];
-node67 [splines=ortho shape = "rect" style = "rounded" label = "549: updateVocabTable" ];
-node73 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "608: saveDocTopics" ];
-node77 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "623: saveTopicWords" ];
-node79 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "647: saveTopicKeys" ];
-node85 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "663: saveTopicPMI" ];
-node91 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "673: saveGraph" ];
-node97 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "697: saveState" ];
-node99 [splines=ortho shape = "rect" style = "rounded,filled" fillcolor = "coral" label = "715: ready" ];
-node100 -> node3 [color="blue" penwidth="2"];
-node100 -> node20 [color="blue" penwidth="2"];
-node14 -> node5 [color="blue" penwidth="2"];
-node14 -> node10 [color="blue" penwidth="2"];
-node20 -> node16 [color="blue" penwidth="2"];
-node20 -> node28;
-node20 -> node32 [color="blue" penwidth="2"];
-node20 -> node34 [color="blue" penwidth="2"];
-node20 -> node47 [color="blue" penwidth="2"];
-node20 -> node67 [color="blue" penwidth="2"];
-node28 -> node22 [color="blue" penwidth="2"];
-node32 -> node26 [color="blue" penwidth="2"];
-node32 -> node49 [color="blue" penwidth="2"];
-node34 -> node10 [color="blue" penwidth="2"];
-node38 -> node5 [color="blue" penwidth="2"];
-node41 -> node26 [color="blue" penwidth="2"];
-node47 -> node38 [color="blue" penwidth="2"];
-node47 -> node41 [color="blue" penwidth="2"];
-node49 -> node34 [color="blue" penwidth="2"];
-node61 -> node51 [color="blue" penwidth="2"];
-node61 -> node55 [color="blue" penwidth="2"];
-node67 -> node55 [color="blue" penwidth="2"];
-node73 -> node5 [color="blue" penwidth="2"];
-node77 -> node5 [color="blue" penwidth="2"];
-node79 -> node28;
-node85 -> node38 [color="blue" penwidth="2"];
-node91 -> node5 [color="blue" penwidth="2"];
-node99 -> node28;
-node99 -> node32 [color="blue" penwidth="2"];
-node99 -> node47 [color="blue" penwidth="2"];
-node99 -> node49 [color="blue" penwidth="2"];
-node99 -> node61 [color="blue" penwidth="2"];
-subgraph clusterjsLDA0{
-node100 node3 node5 node10 node14 node16 node20 node22 node26 node28 node32 node34 node38 node41 node47 node49 node51 node55 node61 node67 node73 node77 node79 node85 node91 node97 node99 node3 node5 node10 node14 node16 node20 node22 node26 node28 node32 node34 node38 node41 node47 node49 node51 node55 node61 node67 node73 node77 node79 node85 node91 node97 node99 ;
-label="jsLDA";
-style=filled;
-color=black;
-graph[style=dotted];
-};
-}
