+++ rubyHTTPServer.rb
@@ -1,0 +1,6 @@
+require 'webrick'
+include WEBrick
+
+s = HTTPServer.new(:Port => 9090,  :DocumentRoot => Dir::pwd)
+trap("INT"){ s.shutdown }
+s.start
